requirements.txt
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
lightgbm>=3.3.5
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
joblib>=1.2.0
openai>=1.0.0
langchain>=0.1.0
jsonschema>=4.17.0
scipy>=1.9.0
langchain-openai>=0.0.1